worked on crud

+++ app/controllers/project_controller.rb
@@ -1,55 +1,22 @@
+
 class ProjectController < ApplicationController
   # returns all projects in the db
-
   get '/projects' do
     projects = Project.all
     projects.to_json
   end
 
-  # gives one project
-
-  get '/project/:id' do
-    project = Project.find(params[:id])
-    project.to_json
-  end
-
-  # gives the names of the members of the project
-  get '/project/members/:id' do
-    project = Project.find_by(id: params[:id].to_i)
-    members = project.users
-    members.to_json
-  end
-
-  # creates a project
-
-  post '/project/create/:id' do
+  post '/project/create' do
     data = JSON.parse(request.body.read)
     begin
-      project = Project.create(data)
-      link = Link.create(user_id: params[:id].to_i, project_id: project.id)
-
+      project = Project.create(title: data['title'], description: data['description'])
       project.to_json
-      link.to_json
-    rescue => e
-      { error: e.message }.to_json
-    end
-  end
-
-  # gets the project id, search a user by name and adds the user as a member of the project
-
-  post '/member/add/:userId/:projectId' do
-    begin
-      user = User.find(params[:userId].to_i)
-      link = Link.create(user_id: user.id, project_id: params[:projectId].to_i)
-
-      link.to_json
     rescue => e
       { error: e.message }.to_json
     end
   end
 
   # updates a project
-
   put '/project/update/:id' do
     data = JSON.parse(request.body.read)
     begin
@@ -61,8 +28,7 @@ class ProjectController < ApplicationController
     end
   end
 
-  # updates status of a prroject
-
+  # updates status of a project
   put '/project/update/status/:id' do
     data = JSON.parse(request.body.read)
     begin
@@ -73,17 +39,39 @@ class ProjectController < ApplicationController
       response(code: 422, data: { error: e.message })
     end
   end
+  # Add a route to add a user to a project
+post '/project/:id/user' do
+  project = Project.find(params[:id])
+  user = User.find(params[:user_id])
+  project.users << user
+  project.to_json(include: :users)
+end
+
+# Add a many-to-many relationship between Project and User
+class Project < ActiveRecord::Base
+  has_and_belongs_to_many :users
+end
+
+class User < ActiveRecord::Base
+  has_and_belongs_to_many :projects
+end
+
 
   # deletes a project
-
   delete '/destroy/:id' do
     begin
       project = Project.find(params[:id].to_i)
-      project.destroy
-
-      response(data: { message: 'project deleted successfully' })
+      if project.nil?
+        response(code: 404, data: { error: 'Project not found' })
+      else
+        project.destroy
+        response(data: { message: 'Project deleted successfully' })
+      end
+    rescue ActiveRecord::RecordNotFound
+      response(code: 404, data: { error: 'Project not found' })
     rescue => e
-      { error: e.message }.to_json
+      response(code: 500, data: { error: e.message })
     end
   end
+  
 end
